Move tools from instru to params
